# Improve SRC performance