Reduce the precision of internal SRC computations to improve resampler performance (by about 20-25%) especially on embedded platforms.

Has no effect on the likelihood of finding a fingerprint match.

Some benchmarks for 120s of audio running on a 1gHz ARM7 CPU:
44.1kHz (before): 2.4s
44.1kHz (after): 2s
176.4kHz (before): 7s
176.4kHz (after): 5.6s