@lalinsky 2017-02-18
Thanks, using integers for the resampling is why I actually included a copy of the old resampling code from FFmpeg in the repo, so having avresample/swresample using floating point was definitely not intentional. 

However, this should be also done for libavresample, not just libswresample, to make sure these are consistent. It's probably better to use the full `internal_sample_fmt` option name for both.

I'm also not sure about the planar format. We want non-planar format out of it in the end. Can you try if it's not even after with `AV_SAMPLE_FMT_S16`?