@agentsim 2017-02-18
`AV_SAMPLE_FMT_S16` isn't supported. I think this is because internally the resampler works on a single channel at a time, so it makes sense that it wouldn't support an interleaved format as the internal format.

I'll do the rename and make the change in libavresample.