@lalinsky 2018-02-16
Yes, but there are still some changes to be made. And it needs to be properly tested with both ffmpeg and libav.