@Seluj78 2019-04-09
@agentsim ping ?